init style2
