+++ composeApp/src/commonMain/kotlin/com/wonddak/portfolio/ui/AboutMe.kt
@@ -13,6 +13,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.wonddak.portfolio.Mode
+import com.wonddak.portfolio.data.iconData
+import com.wonddak.portfolio.data.profile
 import com.wonddak.portfolio.model.IconData
 import com.wonddak.portfolio.model.ProfileData
 import org.jetbrains.compose.ui.tooling.preview.Preview
@@ -28,31 +30,7 @@ fun AboutMe(
     modifier: Modifier,
     mode: Mode,
 ) {
-    val profile = ProfileData(
-        Res.drawable.profile
-    )
-    val iconData = listOf(
-        IconData(
-            Res.drawable.github,
-            "https://github.com/jmseb3",
-            Color(0xff181717)
-        ),
-        IconData(
-            Res.drawable.velog,
-            "https://velog.io/@jmseb3/posts",
-            Color(0xff20C997)
-        ),
-        IconData(
-            Res.drawable.googleplay,
-            "https://play.google.com/store/apps/developer?id=WonDDak",
-            Color(0xff78C257)
-        ),
-        IconData(
-            Res.drawable.appstore,
-            "https://apps.apple.com/kr/developer/jung-wonhee/id1723807756",
-            Color(0xff0D96F6)
-        )
-    )
+
     if (mode == Mode.Desktop) {
         Column(
             modifier,
