+++ composeApp/src/commonMain/kotlin/com/wonddak/portfolio/ui/AboutProject.kt
@@ -12,6 +12,7 @@ import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -21,6 +22,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.wonddak.portfolio.Mode
+import com.wonddak.portfolio.data.projectList
 import com.wonddak.portfolio.model.ProjectData
 import com.wonddak.portfolio.model.ProjectType
 
@@ -53,26 +55,8 @@ fun AboutProject(
             columns = GridCells.Adaptive(minSize = 128.dp)
         ) {
             items(projectList.filter { it.type == nowType }) { item ->
-                Text(item.title)
+                TextButton(onClick = {}){item.title}
             }
         }
     }
 }
-
-val projectList = setOf(
-    ProjectData(
-        1,
-        ProjectType.App,
-        "시원한 수학박사"
-    ),
-    ProjectData(
-        2,
-        ProjectType.Library,
-        "SImp jwt parser"
-    ),
-    ProjectData(
-        3,
-        ProjectType.Plugin,
-        "Font Helper"
-    )
-)
